Дописал старые тесты
