+++ LessonsSwiftUITests/Tests/LoginTextFieldTapDisplaysKeyboardTest.swift
@@ -1,11 +1,13 @@
 import XCTest
 
-final class TapElementFieldLoginTest: BaseTest {
+final class LoginTextFieldTapDisplaysKeyboardTest: BaseTest {
     func testLessonSwift() throws {
         startTests()
         step("Тап на поле 'Логин'") {
             PersonViewControllerScreen.personLogin.tap()
             step("Появилась клавиатура для ввода в поле") {
+                let keyboard = XCUIApplication().keyboards.firstMatch
+                XCTAssertTrue(keyboard.waitForExistence(timeout: 2), "Клавиатура не появилась")
             }
         }
     }
